OSX, C, refact: move sys functions in sys.c
