+++ MANIFEST.in
@@ -97,6 +97,8 @@ include psutil/arch/osx/mem.c
 include psutil/arch/osx/mem.h
 include psutil/arch/osx/process_info.c
 include psutil/arch/osx/process_info.h
+include psutil/arch/osx/sys.c
+include psutil/arch/osx/sys.h
 include psutil/arch/solaris/environ.c
 include psutil/arch/solaris/environ.h
 include psutil/arch/solaris/v10/ifaddrs.c
