+++ psutil/_psutil_osx.c
@@ -40,6 +40,7 @@
 #include "arch/osx/process_info.h"
 #include "arch/osx/cpu.h"
 #include "arch/osx/mem.h"
+#include "arch/osx/sys.h"
 
 
 #define PSUTIL_TV2DOUBLE(t) ((t).tv_sec + (t).tv_usec / 1000000.0)
@@ -540,25 +541,6 @@ error:
             PyErr_WarnEx(PyExc_RuntimeWarning, "vm_deallocate() failed", 2);
     }
     return NULL;
-}
-
-
-/*
- * Return a Python float indicating the system boot time expressed in
- * seconds since the epoch.
- */
-static PyObject *
-psutil_boot_time(PyObject *self, PyObject *args) {
-    // fetch sysctl "kern.boottime"
-    static int request[2] = { CTL_KERN, KERN_BOOTTIME };
-    struct timeval result;
-    size_t result_len = sizeof result;
-    time_t boot_time = 0;
-
-    if (sysctl(request, 2, &result, &result_len, NULL, 0) == -1)
-        return PyErr_SetFromErrno(PyExc_OSError);
-    boot_time = result.tv_sec;
-    return Py_BuildValue("f", (float)boot_time);
 }
 
 
@@ -1410,67 +1392,6 @@ psutil_disk_io_counters(PyObject *self, PyObject *args) {
 error:
     Py_XDECREF(py_disk_info);
     Py_DECREF(py_retdict);
-    return NULL;
-}
-
-
-/*
- * Return currently connected users as a list of tuples.
- */
-static PyObject *
-psutil_users(PyObject *self, PyObject *args) {
-    struct utmpx *utx;
-    PyObject *py_username = NULL;
-    PyObject *py_tty = NULL;
-    PyObject *py_hostname = NULL;
-    PyObject *py_tuple = NULL;
-    PyObject *py_retlist = PyList_New(0);
-
-    if (py_retlist == NULL)
-        return NULL;
-    while ((utx = getutxent()) != NULL) {
-        if (utx->ut_type != USER_PROCESS)
-            continue;
-        py_username = PyUnicode_DecodeFSDefault(utx->ut_user);
-        if (! py_username)
-            goto error;
-        py_tty = PyUnicode_DecodeFSDefault(utx->ut_line);
-        if (! py_tty)
-            goto error;
-        py_hostname = PyUnicode_DecodeFSDefault(utx->ut_host);
-        if (! py_hostname)
-            goto error;
-        py_tuple = Py_BuildValue(
-            "(OOOdi)",
-            py_username,              // username
-            py_tty,                   // tty
-            py_hostname,              // hostname
-            (double)utx->ut_tv.tv_sec, // start time
-            utx->ut_pid               // process id
-        );
-        if (!py_tuple) {
-            endutxent();
-            goto error;
-        }
-        if (PyList_Append(py_retlist, py_tuple)) {
-            endutxent();
-            goto error;
-        }
-        Py_CLEAR(py_username);
-        Py_CLEAR(py_tty);
-        Py_CLEAR(py_hostname);
-        Py_CLEAR(py_tuple);
-    }
-
-    endutxent();
-    return py_retlist;
-
-error:
-    Py_XDECREF(py_username);
-    Py_XDECREF(py_tty);
-    Py_XDECREF(py_hostname);
-    Py_XDECREF(py_tuple);
-    Py_DECREF(py_retlist);
     return NULL;
 }
