+++ setup.py
@@ -244,6 +244,7 @@ elif MACOS:
             'psutil/arch/osx/process_info.c',
             'psutil/arch/osx/cpu.c',
             'psutil/arch/osx/mem.c',
+            'psutil/arch/osx/sys.c',
         ],
         define_macros=macros,
         extra_link_args=[
